requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
SQLAlchemy==2.0.40
gunicorn==21.2.0
psycopg2-binary==2.9.7  # For PostgreSQL in production
openai==1.2.3  # For ChatGPT integration
Werkzeug==2.3.7
Jinja2==3.1.2
pytest==7.4.0  # For testing
requests==2.31.0
MarkupSafe==2.1.3
itsdangerous==2.1.2
Blinker==1.6.2
click==8.1.7
nltk==3.8.1

# Production dependencies
sentry-sdk==1.37.1  # Error tracking
psycopg2-pool==1.1  # Connection pooling for PostgreSQL
